Corrected the NavigationAction.back specs
When attempting to pass `Navigation.back` a null key I was getting errors that said a valid key is required. 
Example:
```javascript
const genericBackAction = NavigationActions.back({
  key: null
});
this.props.navigation.dispatch(genericBackAction);
```
This error contradicts the documentation that was telling me the key was an optional parameter. 
I don't think requiring a key is necessary and a null key should be supported by the library. I may post a PR to the library with the suggested change and implementation. In the meantime, this documentation fix should keep others from hitting this issue.